Simplify the integrate with RSS example
