--- how-to/integrate-with-rss/client/src/integrations/rss/index.ts
@@ -1,5 +1,0 @@
-import { RssIntegrationProvider } from "./integration-provider";
-
-export const entryPoints: { [id: string]: RssIntegrationProvider } = {
-	integrations: new RssIntegrationProvider()
-};
