+++ how-to/integrate-with-rss/public/manifest.fin.json
@@ -41,41 +41,28 @@
 		"target": ["desktop", "start-menu"]
 	},
 	"customSettings": {
-		"homeProvider": {
-			"id": "integrate-with-rss",
-			"title": "Integrate with RSS",
-			"icon": "http://localhost:8080/favicon.ico"
-		},
-		"integrationProvider": {
-			"modules": [
+		"rss": {
+			"icons": {
+				"rss": "http://localhost:8080/icons/rss.svg",
+				"rssEntry": "http://localhost:8080/icons/rss-entry.svg"
+			},
+			"proxyUrl": "http://localhost:8080/proxy",
+			"feedView": "http://localhost:8080/views/rss-feed.html",
+			"feeds": [
 				{
-					"id": "rss",
-					"icon": "http://localhost:8080/favicon.ico",
-					"title": "RSS Provider",
-					"enabled": true,
-					"url": "http://localhost:8080/js/integrations/rss.bundle.js",
-					"data": {
-						"rootUrl": "http://localhost:8080/",
-						"proxyUrl": "http://localhost:8080/proxy",
-						"feedView": "http://localhost:8080/views/rss-feed.html",
-						"feeds": [
-							{
-								"id": "openfin-tech",
-								"url": "https://www.youtube.com/feeds/videos.xml?channel_id=UC61HthWWPCif1wrYNafuyqQ"
-							},
-							{
-								"id": "openfin-versions",
-								"url": "https://cdn.openfin.co/release/meta/rss"
-							},
-							{
-								"id": "youtube-trending",
-								"url": "https://www.youtube.com/feeds/videos.xml?playlist_id=PLrEnWoR732-BHrPp_Pm8_VleD68f9s14-"
-							}
-						],
-						"pollingInterval": 60
-					}
+					"id": "openfin-tech",
+					"url": "https://www.youtube.com/feeds/videos.xml?channel_id=UC61HthWWPCif1wrYNafuyqQ"
+				},
+				{
+					"id": "openfin-versions",
+					"url": "https://cdn.openfin.co/release/meta/rss"
+				},
+				{
+					"id": "youtube-trending",
+					"url": "https://www.youtube.com/feeds/videos.xml?playlist_id=PLrEnWoR732-BHrPp_Pm8_VleD68f9s14-"
 				}
-			]
+			],
+			"pollingInterval": 60
 		}
 	}
 }
