+++ how-to/integrate-with-rss/public/views/rss-feed.html
@@ -1,13 +1,13 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
 	<head>
 		<meta charset="utf-8" />
 		<meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-		<link rel="icon" href="/assets/rss.svg" />
+		<link rel="icon" href="/icons/rss.svg" />
 		<title>RSS Feed</title>
 		<meta name="description" content="" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<script src="./rss-feed.js"></script>
+		<script src="../js/rss-feed.bundle.js"></script>
 		<link rel="stylesheet" href="/common/style/app.css" />
 	</head>
 	<body class="col fill gap20">
@@ -16,7 +16,7 @@
 				<h1 id="feed-title">RSS Feed</h1>
 			</div>
 			<div class="row middle gap10">
-				<image src="../assets/rss.svg" alt="RSS" height="40px"></image>
+				<image src="../icons/rss.svg" alt="RSS" height="40px"></image>
 			</div>
 		</header>
 		<main class="fill col scroll-vertical">
